Logic for displaying notification template properties has been written. Changed file format to jsx
