+++ EventsExpress/ClientApp/src/components/notification-template/index.jsx
