+++ EventsExpress/ClientApp/src/components/notification-template/notification-template-item.jsx
